+++ tests/test_cell_readers.py
@@ -525,6 +525,19 @@ def test_make_summary(cellpy_data_instance, parameters):
     assert s2.iloc[5, 3] == s1.iloc[5, 3]
 
 
+def test_make_summary_new_version(cellpy_data_instance, parameters):
+    cellpy_data_instance.from_raw(parameters.res_file_path)
+    cellpy_data_instance.set_mass(1.0)
+    cellpy_data_instance.make_summary()
+    s1 = cellpy_data_instance.cells[0].summary
+    s2 = cellpy_data_instance.cell.summary
+    s3 = cellpy_data_instance.get_summary()
+    assert s1.columns.tolist() == s2.columns.tolist()
+    assert s2.columns.tolist() == s3.columns.tolist()
+    assert s2.iloc[:, 3].size == 18
+    assert s2.iloc[5, 3] == s1.iloc[5, 3]
+
+
 def test_make_summary_with_c_rate(cellpy_data_instance, parameters):
     cellpy_data_instance.from_raw(parameters.res_file_path)
     cellpy_data_instance.set_mass(1.0)